Add Yargs Builder For Command Argument Optons
Add Yargs Builder argument for when a command argument needs options for
the user to fill out to use that command, for example, add needs to be a
string and is requested. Furthermore, by adding a describe the option
can now appear in the help for add when using `add --help`.