Add `yargs.parse()` Print to Command Line
Add `yargs.parse()` print to command line as there needs to be a way to
tell Yargs when to print the command to the command line so instead of
using `console.log` for this it has its function that is not print the
argument to the command line and confuse the user. Furthermore, Parse
also gathers all of the commands and other Yarge functions and parses
them so that they can use in the command line.