Add Playground & JSON JavaScript File to Play With
Add the playground which is a folder for items not part of the main apps
but something the course needs to show how some parts of JavaScript and
Node work. Add a JSON JavaScript file to show how two of the most useful
functions stringify which turns a JavaScript object into a JSON object
and parse which turns a JSON object into a JavaScript object.