Add GuestList to Event, Show How Func Use `this`
Add the GuestList to the Event object to show how both standard and
arrow functions handle `this`. Standard function bind this to themselves
so the developer can use it for closures and other references in that
function. While Arrow function does not find this so it can access its
parent `this`.