Add Asynchronous Example to Weather App
Add an example for asynchronous to the Weather App to show how
JavaScript uses asynchronous to run some code after the code is first to
run. Also, the Weather app is the next app in the course which is an API
and a frontend to show how Node work with the browser and third-party
APIs.