Add Forecast Function For Chain Callback Functions
Add the function forecast to geocode to chain callbacks together so that
the geocode gets the location data and sends them to the forecast API
thus get the weather for that location.

Chain Callback is when one callback calls another callback and gives
more context to the next callback so that they can make other function
calls with the data.