Add Handlebar Partials to Views
Add some handlebar partials which are snippets of HTML that can be
loaded into an HTML page and be a template where a single change can
affect multiple pages are the same time.

Also, Nodemon needs to use `-e js,hbs` in its command line so that so
that it can auto rebuild when '.hbs' files are changed.