Add 404 Route to Express
Add a 404 route to Express when no longer routes, are found. A 404 route
needs to be last because once a route is found Express is going stop
looking for other routes since the route that is in the URL is found.
If not routes are found then a 404 route, is rendered which is why it
only has the wildcard character.