Add Example For Object Default Params in Function
Add an example for the object default params, in the function, as an
object default params can use two different ways to assign a default
params. One is to have an object assigned to the whole destructed prams;
however, this course the values in that object to be undefined since the
object assign is an empty object. The second object default params are
to assign the destructed params directly which is preferred since to
control what is the value being display to the user.