Add Nodemon Module as Dev Dependence in Web Server
Add the Nodemon module as dev dependence in Web Server as it is needed
to develop locally. Having a global module is not recommend since the
developer may use that global module in a Node script that other
developers do not have thus making the project complex and having them
needing to install other modules outside of the project that is why to
add dependencies, are needed for projects.