Add Validator Package For More Complex Validations
Add the Validator package for more complex validations so that we do not
have to write out all of the possible string validations needed to make
sure that the string data is following business logic. Mongoose uses
simple validations for strings, and as a result, it does not have
validations for if an email format is correct or if an IP format is also
correct.

Ref:

https://www.npmjs.com/package/validator