Add Example For Promises Chaining
Add Example For Promises Chaining, which is a way to async sequence
functions so that once one async function, is done. The results, of
that, is pass to another async function for more processing an example
of this is getting the geocodes from a service then getting the weather,
from another service, based on what the geocode return and then return
the results.