Add Example For Await in Playground
Add example for Await in Playground which must is in a function that is
using Async so that the developer can then add Await to any function
called that is returning a Promise this makes the async easier to read
as we now know which functions are Async and is going to wait for the
results to come back. Furthermore, you can chain Await functions by
getting a value from a function call another function and use that value
in an argument for the other function.