Add Example of Bcrypt in Task Manager Index
Add an example of Bcrypt, in Task Manager index page, to show how to use
it. Also, the 8 means the number of rounds the hashing should go, too
many rounds slows down the application, on the other hand, too few makes
hash easy to crack. So eight is the number required by the creator of
Bcrypt.