Update User Model to Create Schema For Middleware
Update User model to create a schema for to use Mongoose middleware
which can take a schema and can perform actions on it before being added
to the Mongoose Model function. Since usually Mongoose can take the
second argument when creating a model and create a schema for us.
However, the developer you can do this themselves and have accurate
control of how to the data used before Mongoose can use it.

Ref:

https://mongoosejs.com/docs/middleware.html