Add Pre Mongoose Middleware
Add Pre Mongoose middleware so any action before an event this is
called. Furthermore, the second argument has to be a normal function
since Mongoose uses `this` which is the schema that is calling this
function. Additionally, there is an argument for that function called
next which must be called at the end of the function to tell Pre to end
the function otherwise the function is going to hang and not save the
item.

The Post is doing the same as Pre but does actions after an event.

Ref:

https://mongoosejs.com/docs/middleware.html