Add Mongoose Methods to User `generateAuthToken`
Add Mongoose methods to the User model for `generateAuthToken` the
generate JWTs after login. Methods allow the developer to add their
custom function as an instance method to a model. Since JWTs are per
user and not create a single JWT for every user, a custom instance
method makes sense in this regard.