Add Tokens Array & Token Object to User Model
Add a tokens array and token object to User model so to track the number
of tokens the user is using as they can be login to different devices so
not have to log out of every device when the logout. The developer can
remove a single token for that device while keeping the user login to
other devices.