Add Example For Express Middleware
Add an example for Express middleware where it sits between the new
request and running the route requested. Where the developer can add
function to either log an event or verify a token then let the route
continue by calling `next()` is provided by Express or cancel the
request if there is an issue.