Add Middleware Folder & Auth For Authiorizes
Add the middleware folder where all Express middle were is located as
that is the pattern when using them since there could be many of them
that could use, in not only all the routes but also an original route.

Also, add Auth to do authorizes to make sure that the user can do what
they are authorized to do.