Switch User Index Route to Show Route
Change the route for User from Index to Show Route as Index route is not
needed since there is no use in having a route where all the users are
displayed. A user should not need to know the list of users, so instead,
considering we know the user already as they were, we can then use that
route to show that users data.