Update `toJSON` to be Serialization For User Model
Update the `toJSON` function to be serialization for the User model
instead of using the custom function as any User instance uses `toJSON`
to get the data from the User model. As a result, we can override that
function to serialization so that any data that the developer needs is
going to get the serialized form since no one needs the password or
token outside of the User model.