Add Virtual Tasks Property For User Model
Add a virtual property Which is called Tasks for the User model which
allows the developer to access any task attach to that user. It is a
property that is in memory that connects models; it is good with it is a
one to many relationships. Furthermore, it uses local field which is the
field that is attached to in this case the user id and the different
field which is the foreign key in the Task table, which is the owner's
field.