Add Multer Upload User Avatar For 56th Challenge
Add the Multer package to upload a user avatar for the 56th challenge as
Multer uses `form-data,` in the request body to create the file in a
Node server of what the user has uploaded. Also, for the single function
which is the HTTP request key so that Multer knows which key to use for
that path or multiple file types.