Update Task Manager `.gitignore` to Reflect State
Update the Task Manager's `.gitignore` to reflect that current state of
the repo meaning that images are, in the database, so there is no needed
to hide them. Config no longer needs a star to as it is in the root
directory, and `node_modules` folder needs to be separated so that the
Task Manager application can stand on its on when deploying to Heroku.