Add Example For Jest Test Using Async / Await
Add example for Jest test using Async / Await, which means not to use
`done` as an argument since Async is returning a promise. As a result,
Jest is going to wait until the promise, is fulfilled, before ending the
test, meaning no need to tell Jest when the async function, is called
it.