Add Mocks For SendGrid Mail in Jest Test
Add the mocks which are functions the replace the actual function for a
test so that they do something different without touch production data.
In this example for SendGrid to not send out an email in Jest test as
each signup and delete user has a to send out an email and since those
test, they go nowhere and cost money to send out.