Add Expect `toEqual` to User Avatar Test
Add the expect `toEqual` to the User Avatar test to find out if one
object is the same as another object since toBe is the same as `===`
which is incorrect for objects as two object two different parts of
memory which `===` considers to be not correct. Instead, `toEquals` goes
more in-depth into an object to see if they are the same.
Expect.any(Object) wants a way to return all of the properties for an
object to find out if an object is the same as what is coming back from
test results.