Add Socket Event Join Room on Server Side
Add the Socket event join(room) on Server side so get two more emit
functions `io.to.emit` and `socket.broadcast.to.emit`. `io.to.emit`
sends to all connections to a given room instead of everyone connect to
the server this way connections can have a private connect without
everyone that is connected, to the server. `socket.broadcast.to.emit`
sends to all connections expect the user that started the broadcast same
as `socket.broadcast.emit`, but local to a room.