Add Auto Scrolling to Chatbox
Add auto-scrolling to the Chatbox so that the scrolling automatically
follows the new message if the user is not manually scrolling throw
history. Since they may be looking at a past message and the chatbox
should not send them back to the newest message, which makes for lousy
user experience.