Add First GraphQL Type Definition or Schema
Add the first type definition also known as a schema which tells what
type the data is going to become since there needs to be a way for
GraphQL to know the type of data that, is used. Also, a bang means that
the type is always returned value since it is a non-nullable field.