Add Arguments For Query Types
Add an argument to a query type to show how to set up arguments with
resolvers which are suitable for querying large datasets or add/edit
data to the database from the frontend.

In the resolver there are four parameters that are attached to each
resolver, they are parent which is the relationship to the query type,
args, for the arguments that the developer has passed in, ctx known as
context getts access to any information that was added-in by the user
earlier on in the application, and info for the information on the query
type.