Add Example of Array Type
Add an example of an array type where a developer can assign an array of
a type using bang at the end for non-nullable as well as add bang to the
type for non-nullable elements for example, `grades: [Int!]!` means
grades cannot be nullable as well as any numbers in grades.

Furthermore, when querying an array, the developer does not need curly
braces as the elements in an array with basic types are not an object.